requirements.txt
PyQt6>=6.7.0
PyQt6-Qt6>=6.7.0
httpx>=0.27.0
requests>=2.32.0
python-docx>=1.1.0
markdown2>=2.5.0
markdownify>=0.13.0
pydub>=0.25.1
audioop-lts>=0.2.1  # Python 3.13+ replacement for removed audioop stdlib module
ffmpeg-python>=0.2.0
kokoro>=0.9.0
TTS>=0.22.0
soundfile>=0.12.1
numpy>=1.26.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
platformdirs>=4.2.0
